XLA CustomCall interface for envpool (#149)
Modern DL frameworks like JAX/Tensorflow allows lowering the neural network part of the loop down to native code. However, currently envpool is pure python, meaning the interaction between the algorithm and the env has to be in python.

With this PR, the neural nets can be jitted together with envpool, which minimizes the dispatching latency.